+++ solstice/src/main/java/dev/equo/ide/IdeHookBranding.java
@@ -13,15 +13,25 @@
  *******************************************************************************/
 package dev.equo.ide;
 
+import com.diffplug.common.swt.os.OS;
 import java.io.File;
 import java.io.IOException;
 import java.util.Arrays;
 import javax.annotation.Nullable;
+import org.eclipse.core.internal.runtime.InternalPlatform;
+import org.eclipse.core.runtime.IProduct;
+import org.eclipse.jface.preference.IPreferenceStore;
+import org.eclipse.jface.window.Window;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.graphics.Image;
 import org.eclipse.swt.graphics.ImageData;
 import org.eclipse.swt.widgets.Display;
 import org.eclipse.swt.widgets.Shell;
+import org.eclipse.ui.internal.ide.IDEInternalPreferences;
+import org.eclipse.ui.internal.ide.IDEWorkbenchPlugin;
+import org.osgi.framework.Bundle;
+import org.osgi.framework.BundleContext;
+import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 @SuppressWarnings("serial")
@@ -52,6 +62,15 @@ public class IdeHookBranding implements IdeHook {
 	}
 
 	class Instantiated implements IdeHookInstantiated {
+		Logger logger = LoggerFactory.getLogger(IdeHookBranding.class);
+
+		@Override
+		public void isClean(boolean isClean) throws Exception {
+			if (OS.getRunning().isMac()) {
+				System.setProperty("com.apple.mrj.application.apple.menu.about.name", title);
+			}
+		}
+
 		private Shell splash;
 
 		private Image loadImage(Display display, File file, String defaultResource) {
@@ -75,6 +94,57 @@ public class IdeHookBranding implements IdeHook {
 		}
 
 		@Override
+		public void afterOsgi(BundleContext context) {
+			var internal = InternalPlatform.getDefault();
+			try {
+				var product = internal.getClass().getDeclaredField("product");
+				product.setAccessible(true);
+				product.set(
+						internal,
+						new IProduct() {
+							@Override
+							public String getApplication() {
+								return title;
+							}
+
+							@Override
+							public String getName() {
+								return title;
+							}
+
+							@Override
+							public String getDescription() {
+								return title;
+							}
+
+							@Override
+							public String getId() {
+								return title;
+							}
+
+							@Override
+							public String getProperty(String key) {
+								return null;
+							}
+
+							@Override
+							public Bundle getDefiningBundle() {
+								return Arrays.stream(context.getBundles())
+										.filter(bundle -> bundle.getSymbolicName().equals("dev.equo.ide"))
+										.findFirst()
+										.get();
+							}
+						});
+
+				IPreferenceStore ps = IDEWorkbenchPlugin.getDefault().getPreferenceStore();
+				ps.setValue(IDEInternalPreferences.SHOW_LOCATION_NAME, false);
+				ps.setValue(IDEInternalPreferences.SHOW_PRODUCT_IN_TITLE, true);
+			} catch (Exception e) {
+				logger.warn("problem defining product", e);
+			}
+		}
+
+		@Override
 		public void afterDisplay(Display display) {
 			var cursor = display.getCursorLocation();
 			var monitors = display.getMonitors();
@@ -89,7 +159,7 @@ public class IdeHookBranding implements IdeHook {
 			int imgX = image.getBounds().width;
 			int imgY = image.getBounds().height;
 			splash = new Shell(display, SWT.NO_TRIM | SWT.ON_TOP);
-			splash.setText("Branding");
+			splash.setText(title);
 			splash.setBounds(
 					bestMonitor.x + (bestMonitor.width - imgX / 2) / 2,
 					bestMonitor.y + (bestMonitor.height - imgY / 2) / 2,
@@ -106,10 +176,19 @@ public class IdeHookBranding implements IdeHook {
 			splash.open();
 			splash.forceActive();
 
-			Display.setAppName(title);
 			while (display.readAndDispatch())
 				// pump the event loop enough to show the branding
 				;
+
+			// now set the application icon
+			Image icon =
+					loadImage(Display.getDefault(), IdeHookBranding.this.icon, "dev/equo/ide/equo_icon.png");
+			var bounds = icon.getBounds();
+			if (bounds.width != bounds.height) {
+				LoggerFactory.getLogger(IdeHookBranding.class)
+						.warn("Icon should be square, but is instead {} by {}", bounds.width, bounds.height);
+			}
+			Window.setDefaultImage(icon);
 		}
 
 		@Override
@@ -117,27 +196,14 @@ public class IdeHookBranding implements IdeHook {
 			splash.dispose();
 			splash = null;
 
-			Display.getDefault()
-					.asyncExec(
-							() -> {
-								var display = Display.getCurrent();
-								if (display == null) {
-									// early shutdown
-									return;
-								}
-								Display.setAppName(title);
-								Image icon =
-										loadImage(
-												Display.getDefault(),
-												IdeHookBranding.this.icon,
-												"dev/equo/ide/equo_icon.png");
-								Shell[] shells = display.getShells();
-								for (var shell : shells) {
-									shell.setText(title);
-									shell.setImage(icon);
-									shell.forceActive();
-								}
-							});
+			// setup the task item
+			var taskBar = Display.getDefault().getSystemTaskBar();
+			if (taskBar != null) {
+				var item = taskBar.getItem(null);
+				if (item != null) {
+					item.setText(title);
+				}
+			}
 		}
 	}
 }